Move synchronous tests to own subfolder
